### 2025-05-09 #8186 @editorialbot
[PRE REVIEW]: OceanLight.jl: A Julia package for calculating irradiance in the upper ocean using Monte Carlo simulations
**Submitting author:** @haoboatlab (<a href="http://orcid.org/0000-0003-4898-1074">郝轩亭 (Hao, Xuanting)</a>)
**Repository:** https://github.com/haoboatlab/OceanLight.jl
**Branch with paper.md** (empty if default branch):
**Version:** v.1.0.0
