### 2025-05-13 9085355
Adding DOI in joss/paper.bib
