+++ joss/paper.bib
@@ -16,7 +16,8 @@
   volume={44},
   number={1},
   year={2006},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1029/2003RG000148}
 }
 
 @book{Mobley:1994,
@@ -36,7 +37,8 @@
   title={Volume scattering functions for selected ocean waters},
   author={Petzold, Theodore J},
   year={1972},
-  publisher={Scripps Institution of Oceanography La Jolla, CA, USA}
+  publisher={Scripps Institution of Oceanography La Jolla, CA, USA},
+  doi = {10.21236/ad0753474}
 }
 
 @article{Hao:2022,
@@ -47,7 +49,8 @@
   number={11},
   pages={e2022GL097769},
   year={2022},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1029/2022GL097769}
 }
 
 @article{Smith:1981,
@@ -58,7 +61,8 @@
   number={2},
   pages={177--184},
   year={1981},
-  publisher={Optica Publishing Group}
+  publisher={Optica Publishing Group},
+  doi = {10.1364/ao.20.000177}
 }
 
 @article{Darecki:2011,
@@ -68,7 +72,8 @@
   volume={116},
   number={C7},
   year={2011},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1029/2011JC007338}
 }
 
 @article{Gernez:2011,
@@ -78,7 +83,8 @@
   volume={116},
   number={C7},
   year={2011},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi = {10.1029/2011JC007156}
 }
 
 @article{Xu:2014,
@@ -89,7 +95,8 @@
   number={5},
   pages={051408--051408},
   year={2014},
-  publisher={Society of Photo-Optical Instrumentation Engineers}
+  publisher={Society of Photo-Optical Instrumentation Engineers},
+  doi = {10.1117/1.OE.53.5.051408}
 }
 
 @phdthesis{Xu:2011,
