### 2025-05-20 4bc0db3
Add Doi for Dickey:2011 in joss .bib
