+++ joss/paper.bib
@@ -6,7 +6,8 @@
   number={4},
   pages={44--49},
   year={2011},
-  publisher={AIP Publishing}
+  publisher={AIP Publishing},
+  doi = {10.1063/1.3580492}
 }
 
 @article{Dickey:2006,
