### 2025-06-25 #8186 @AnjaliSandip
> Hi [@AnjaliSandip](https://github.com/AnjaliSandip) thanks for handling our paper. Below is a list of reviewers that I suggest:
>
>     1. Yidongfang Si, [y_si@mit.edu](mailto:y_si@mit.edu), Dr. Si is an expert in geophysical fluid dynamics and she combines theoretical approaches and numerical simulations to investigate fundamental problems in physical oceanography.
>
>     2. Bia Villas Bôas ([biavillasboas](https://github.com/biavillasboas)), [villasboas@mines.edu](mailto:villasboas@mines.edu), Dr. Villas Bôas is a physical oceanographer who combines observations, theory, and modeling to understand how coupled air-sea interactions affect the environment.
>
>     3. An Wang, [awang38@stevens.edu](mailto:awang38@stevens.edu), Dr. Wang is an expert in ocean wave applications and the fundamental and critical problems in free surface flows.
>
>     4. Gregory L. Wagner ([glwagner](https://github.com/glwagner)), [gregory.leclaire.wagner@gmail.com](mailto:gregory.leclaire.wagner@gmail.com), Gregory L. Wagner, Dr. Wagner is one of the main developers of Oceananigans, a ocean model written in Julia.
>
>     5. iuryt (https://github.com/iuryt)

Thank you!
