### 2025-06-26 #8186 @AnjaliSandip
👋 @biavillasboas, @glwagner, and @iuryt, would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:[ https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html).
