### 2025-07-03 #8186 @glwagner
> > Hi, I'd be happy to review this paper. What is the due date?
>
> Preferably, within 3 weeks.

I'll need a bit more time since I will be traveling and on vacation for much of the next 3 weeks. Hope that's ok!
