### 2025-07-30 c9dc168
Improve README description
Also fix a typo.
