### 2025-08-31 971ad73
Simplify test files
