### 2025-08-31 haoboatlab/OceanLight.jl#9 @glwagner
Guidelines for contributing to this package?
I could not find guidelines for contributing to this package (when to open an issue, how to communicate issues with the code, and any style guidelines or coding practices that should be taken into account in a pull request). It would be nice if even a simple comment could be added to the README and/or the documentation.

x-ref: https://github.com/openjournals/joss-reviews/issues/8537
