+++ docs/src/module/MonteCarlo.md
@@ -7,6 +7,7 @@ The basis of Monte Carlo method lies in the idea that, if we know the probabilit
 *Illustration of three photons that are being emitted and travel inside the water medium*[^1]
 
 As mentioned above, the separate event that happens for each photon is consisted of travelling some distance in the medium, aborbing, and scattering. All of the events are determined by some probability. Hence, we define the probability distribution of a random number $p_{\Re}$[^1]
+
 ```math
 p_{\Re} (\Re) \equiv \begin{cases}
 1 & \quad 0 \leq \Re \leq 1, \\
@@ -16,14 +17,16 @@ p_{\Re} (\Re) \equiv \begin{cases}
 
 ## The optical path length
 
-After the photon's interaction with the medium or being transmitted into the water, we calculate for the distance the photon would travel before interact with the medium agian. The distance $r$ is defined by the equation below, when $c$ is the beam attenuated coefficient or the addition of absorbtance coefficient $a$ and scatterance coefficient $b$. [^1]
+After the photon's interaction with the medium or being transmitted into the water, we calculate for the distance the photon would travel before interact with the medium again.
+The distance $r$ is defined by the equation below, when $c$ is the beam attenuated coefficient or the addition of absorbtance coefficient $a$ and scatterance coefficient $b$. [^1]
 
 ```math
 r=-\frac{1}{c}\ln{\Re}
 ``` 
 ## Sampling photon interaction type
 
-Once, the photon travels for distance $r$, we, then, determine how it interacts with the medium: whether a photon will be absorbed or scattered. This can be done by drawing a random number $\Re$ and compare with the albedo of single scattering $\omega_{0} = \frac{b}{c}$, if 
+Once the photon travels for distance $r$, we then determine how it interacts with the medium; e.g. whether a photon will be absorbed or scattered.
+This can be done by drawing a random number $\Re$ and compare with the albedo of single scattering $\omega_{0} = \frac{b}{c}$, if 
 $\Re\leq\omega_{0}$, the photon is being absorbed, and if $\Re\geq\omega_{0}$, the photon is being scattered. [^1]
 
 ## Sampling scattering direction
@@ -33,7 +36,7 @@ Azimuthal angle ``\varphi`` is uniformly distribute between 0 and ``2\pi`` hence
 ```math
 \varphi = 2\pi\Re
 ``` 
-The angle between new trajectory and the direction before scattering ``\theta`` is defined by the cumulative distribution, based on the experiment conducted by the Petzold(1972). [^2]
+The angle between new trajectory and the direction before scattering ``\theta`` is defined by the cumulative distribution, based on the experiment conducted by the Petzold (1972). [^2]
 
 ![CDF](https://raw.githubusercontent.com/haoboatlab/OceanLight.jl/main/docs/src/assets/CDF_scattering_angle.png)
 
