### 2025-08-31 haoboatlab/OceanLight.jl#11 @glwagner
Statement of need?
The README states:

> OceanLight.jl calculates the downwelling irradiance field in the upper ocean. It uses the Monte Carlo method to simulate the trajectory of photons, starting from the refraction in the air-water interface, to some specific depth underneath the water body.

The JOSS draft contains an excellent statement of need:

> Optical oceanography concerns all aspects of light and its interaction with seawater, which are crucial for addressing problems related to physical, biological, and chemical oceanographic processes, such as phytoplankton photosynthesis, biogeochemical cycles, and climate change. However, due to the complex interaction between light and free-surface wave geometry, the irradiance distribution can be highly variable (Darecki et al., 2011; Gernez et al., 2011), making it difficult to obtain analytical solutions.

My request is merely that a paragraph like this (or identical) appear also in the README (and ideally also in the documentation).

x-ref: https://github.com/openjournals/joss-reviews/issues/8537
