### 2025-08-31 haoboatlab/OceanLight.jl#12 @glwagner
Other similar packages?
I am curious whether the authors are aware of similar packages --- in any programming language. For example, one package that is vaguely related is [`OceanBioME`](https://github.com/OceanBioME/OceanBioME.jl), which as far as I know also contains (crude, bulk) models for the absorption of radiation by seawater. It's nice to explain both the links and differences between these packages to help users pinpoint the use case for this software.

cc @jagoosw

xref https://github.com/openjournals/joss-reviews/issues/8537
