### 2025-08-31 haoboatlab/OceanLight.jl#13 @glwagner
Is this package parallelized with MPI? With multithreading?
The JOSS draft states:

> OceanLight supports parallel processing via MPI, distributing photons evenly across CPU threads.

Parallelization with MPI is distinct from parallelization with threads. MPI launches many processes at once, which must communicate with MPI directives. Threads use a single program paradigm, where a "main program" can control all of the threads.

This package does not contain MPI.jl as a dependency:

https://github.com/haoboatlab/OceanLight.jl/blob/2ef55de6188886b9622cf3c57e8e564801e40a5b/Project.toml#L6-L14

therefore I think it is unlikely that the package is parallelized with MPI.

I also cannot find any reference to `Threads.@threads` or other parallel constructions. I might be missing something -- but is it certain that multithreading parallelization is actually used in this package?

xref: https://github.com/openjournals/joss-reviews/issues/8537
