### 2025-09-02 c54ab22
Fix typo in 'Center.md' and adjust formatting for clarity
