### 2025-09-02 f5f3de3
fix typos, change organization of tutorial and make plot more clear, with no blank spaces.
