+++ docs/src/QuickStart/Center.md
@@ -50,7 +50,7 @@ nxeta = 512                 # Number of surface elevation grid point in x direct
 nyeta = 512                 # Number of surface elevation grid point in y direction
 pex = 2*pi/20.0             # Lowest wavenumber that can be captured during the derivative of surface elevation in y direction
 ```
-**NOTE:** `num` must be set to a constant value of 31 (number of angle measurement in Kirk, 1981 [^2]). `kbc` accepts only binary values (0 or 1), representing periodic boundary conditions. In contrast to grid spacing `dz` in z-direction, , the grid spacings in the x- and y-directions (`dx` and `dy`) are calculated automatically by `OceanLight.readparams`,  using the formulas $dx \times nxe  = \frac{2\pi}{pex}$, and $dy \times nye = \frac{2\pi}{pey}$. For more detail on all parameters  used, see [`Simulation parameters`](@ref simulation_parameters). 
+**NOTE:** `num` must be set to a constant value of 31 (number of angle measurement in Kirk, 1981 [^2]). `kbc` accepts only binary values (0 or 1), representing periodic boundary conditions. In contrast to grid spacing `dz` in z-direction, the grid spacings in the x- and y-directions (`dx` and `dy`) are calculated automatically by `OceanLight.readparams`,  using the formulas $dx \times nxe  = \frac{2\pi}{pex}$, and $dy \times nye = \frac{2\pi}{pey}$. For more detail on all parameters  used, see [`Simulation parameters`](@ref simulation_parameters). 
 
 To create a input variable file suitable for this package, user can either create a new file in `.yml` format, copy, and paste the code block above, or using a build-in function `OceanLight.writeparams` to automate the process. 
 
@@ -74,9 +74,7 @@ p = OceanLight.readparams()
 
 Before the simulation, user needs to declare and initialize all parameters and their dimensions. 
 
-### Air-Water Interaction
-
-During the light refraction between two mediums calculation, `OceanLight` requires the surface elevation attribution: $\eta$; surface elevation, $\eta_{x}$; partial derivative of $\eta$ in x direction, and $\eta_{y}$; partial derivative of $\eta$ in y direction. All surface elevation distribution, as specified above, need to have the same dimension with the incoming photons' grid. Hence, 
+During the light refraction between two mediums (air and water) calculation, `OceanLight` requires the surface elevation attribution: $\eta$; surface elevation, $\eta_{x}$; partial derivative of $\eta$ in x direction, and $\eta_{y}$; partial derivative of $\eta$ in y direction. All surface elevation distribution, as specified above, need to have the same dimension with the incoming photons' grid. Hence, 
 
 ```@example Center
 η = zeros(p.nxs,p.nys)
@@ -94,7 +92,17 @@ zpb = zeros(p.nxp,p.nyp);
 fres = zeros(p.nxp,p.nyp);
 ```
 
-### Monte Carlo simulation
+During the air-water interaction process, OceanLight simulates the photons transfer directly downward from the air side, interacts with the water surface, and transfer down into water medium. 
+
[…]
