### 2025-09-02 07a534f
Enhance plot clarity by adjusting axis limits, add colorbar and adding transparency to lines in heatmaps
