### 2025-09-15 06ce16f
Add brief statement of need to Readme, index.md
Adding a brief statement of need into README.md and src/index.md file.
Include adding the reference sections into src/index.md file.
