### 2025-09-15 365f821
Add "running first model" section on Readme.md
Adding a section Running your first model, containing the code used to generate center photons in Readme.md file
