+++ README.md
@@ -18,6 +18,7 @@ OceanLight requires Julia software. To do so,
 2. Launch Julia and type 
 
 ```Julia
+cd(mktempdir())
 using Pkg
 Pkg.add("OceanLight")
 ```
@@ -27,6 +28,147 @@ After installing, verify that OceanLight works as intended by:
 ```Julia
 Pkg.test("OceanLight")
 ```
+
+## Running your first model 
+
+As a simple example, let’s calculate the downwelling irradiance field when the surface is completely flat and a total of 10,000,000 photons are focused at a single point in the center.
+
+```Julia
+# irradiance
+nz = 200                    # Number of total grid point in z direction
+dz = 1                      # Physical length of grid spacing in z direction
+nxe = 512                   # Number of grid point of the calculation grid in x direction  
+nye = 512                   # Number of grid point of the calculation grid in y direction  
+num = 31                    # Constant value (number of angle measurement in Kirk,1981)
+ztop = 10                   # Number of grid point in air phase in z direction
+# photon
+nphoton = 10000000          # Number of photons being generated at each grid point 
+kr = 10                     # Dummy variable (in developement, not being used)                    
+nxp = 512                   # Number of grid points in x direction where photon can be emitted
+kbc = 0                     # Binary value 0 and 1 depending on Boundary condition being implemented 
+b = 0.0031                  # Scattering coefficient 
+nyp = 512                   # Number of grid points in x direction where photon can be emitted
+a = 0.0196                  # Absorbtance coefficient
+# wave
+pey = 2*pi/20.0             # Lowest wavenumber that can be captured during the derivative of surface elevation in x direction
+nxeta = 512                 # Number of surface elevation grid point in x direction
+nyeta = 512                 # Number of surface elevation grid point in y direction
+pex = 2*pi/20.0             # Lowest wavenumber that can be captured during the derivative of surface elevation in y direction
+
+data=Dict("irradiance"=>Dict("nxe"=>nxe,"nye"=>nye,"nz"=>nz,"dz"=>dz,"ztop"=>ztop,"num"=>num),
+            "wave"=>Dict("pex"=>pex,"pey"=>pey,"nxeta"=>nxeta,"nyeta"=>nyeta),
+            "photon"=>Dict("nxp"=>nxp,"nyp"=>nyp,"nphoton"=>nphoton,"a"=>a,"b"=>b,"kr"=>kr,"kbc"=>kbc))
+
+OceanLight.writeparams(data)
+p = OceanLight.readparams()
+
+η = zeros(p.nxs,p.nys)
+ηx = zeros(p.nxs,p.nys)
+ηy = zeros(p.nxs,p.nys)
+xpb = zeros(p.nxp,p.nyp);
+ypb = zeros(p.nxp,p.nyp);
+zpb = zeros(p.nxp,p.nyp);
+θ = zeros(p.nxp,p.nyp);
+ϕ = zeros(p.nxp,p.nyp);
+fres = zeros(p.nxp,p.nyp);
+ed = zeros(p.nx, p.ny, p.nz)
+esol = zeros(p.num, p.nz)
[…]
