### 2025-09-15 d0d80b9
Fix minor detail ReadMe/ Add quick model src/index.md
In README.md file, fixing the code in "Running your first model" section, by adding importing code part. Adding new images 1e7 photons in /docs/src/asset. Adding the new reference section. In src/index.md adding the whole section of "Running your first model".
