+++ docs/src/index.md
@@ -164,9 +164,8 @@ plot(p1, p2, p3, layout=l,
      titleloc=:left, titlefont=font(8),
      left_margin=10mm, right_margin=10mm)
 ```
-<img  src="https://raw.githubusercontent.com/haoboatlab/OceanLight.jl/main/docs/src/assets/Center1e7.png" width="600" align="center">
-
 </details>
+<img  src="https://raw.githubusercontent.com/haoboatlab/OceanLight.jl/main/docs/src/assets/center1e7.png" width="600" align="center">
 For a complete guide with details on each function and step, see the [OceanLight's Documentation](https://haoboatlab.github.io/OceanLight.jl/dev/QuickStart/Center/). 
 
 ## Result
