### 2025-09-16 475b58d
Add contribution and contribute.md
Adding contribution section in README.md and adding the docs/src/contribute.md pages
