+++ docs/Project.toml
@@ -1,2 +1,4 @@
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
+OceanLight = "ff88e13d-0366-4369-baeb-7c3fb86ee06b"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
