+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"module/refraction/#Refraction","page":"Refraction","title":"Refraction","text":"","category":"section"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"To calculate the refraction of the photons between two medium: air and water, OceanLight.jl expands the Snell's Law into three dimension spherical coordinate, with theta_t denotes the transmission angle beneath the surface, and varphi denotes the azimuth angle. In addition to the photons' path, OceanLight.jl calculates the fraction of light that is being reflected back to the air side, via Fresnel reflectance. ","category":"page"},{"location":"module/refraction/#Snell's-Law","page":"Refraction","title":"Snell's Law","text":"","category":"section"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"The angle of reflection describes through the vector analysis, [1]","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"theta_r = cos^-1hatxicdothatn","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"When our incoming photon is directly downward: hatxi=beginbmatrix001endbmatrix and the normal vector can be defined by hatn=dfrac1sqrt1+left(eta_xright)^2+left(eta_yright)^2left(-eta_xhati-eta_yhatj+hatkright) When, eta_x is the partial derivative the partial derivative of eta with respect to x: dfracpartialetapartial x and eta_y is the partial derivative the partial derivative of eta with respect to y: dfracpartialetapartial y","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"Therefore, the angle of reflection, in this module, can be described by ","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"theta_r = cos^-1left(frac1sqrt1+left(eta_xright)^2+left(eta_yright)^2right)","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"The angle of transmission can be described by Snell's Law,","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"theta_t = sin^-1left(frac1n_wsintheta_rright)","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"Substitute the theta_r that we found above. ","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"theta_t = sin^-1left(frac1n_wsqrtfracleft(eta_xright)^2+left(eta_yright)^21+left(eta_xright)^2+left(eta_yright)^2right)","category":"page"},{"location":"module/refraction/#Fresnel-reflectance","page":"Refraction","title":"Fresnel reflectance","text":"","category":"section"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"In our package, we calculate the energy proportion of the light ray that being transmitted to the water, but first we identify the amplitude transmission coefficient or the ratio between electric field amplitude of the transmitted light ray to the intirial light ray. [2]","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"t_perp = left(fracE_tE_0right)_perp = frac2sin(theta_t)cos(theta_r)sin(theta_t+theta_r)","category":"page"},{"location":"module/refraction/","page":"Refraction","title":"Refraction","text":"t_parallel = l
[…]
