### 2025-09-19 3b684bc
Add quick tutorial on index.md
Adding "running first model" section on the docs/src/index.md file. Also change the subsection title ##result to ##Gallery  on Readme.md and index.md
