### 2025-09-19 aed88b7
Fixing typo in docs/src/index.md
deleting the detail and summary bracket in the code for plotting, and fix the image path.
