### 2025-09-19 haoboatlab/OceanLight.jl#11 @tayT0T
Thank you for the comment. We agree that including a statement of need in the README and on the front page is helpful. We have added it in commit 06ce16f.
