### 2025-09-19 haoboatlab/OceanLight.jl#9 @tayT0T
Thank you for the comment. We agree and have revised our documentation accordingly. Following the examples of Oceananigans.jl, OceanBioME.jl, and WaterLiLy.jl, we added a brief Contributing section to the README pages, as well as a full Contributor Guide in the documentation (commit 475b58d).
