### 2025-09-19 haoboatlab/OceanLight.jl#13 @tayT0T
Thank you for the comment. You are correct that the package dependencies do not use MPI.jl; therefore, OceanLight.jl does not directly incorporate MPI.jl into its functions. Instead, OceanLight.jl is designed so that, within a specific workflow, users can integrate [MPI.jl](http://mpi.jl/), ie., by assigning how many photons each CPU will calculate (given initial position and direction) and then combining the irradiance fields across all CPU.
We agree that, without an example or tutorial, the statement in the JOSS draft may be confusing. To address this, we are preparing a new page under docs/src/QuickStart, which will include a tutorial and example on using MPI.jl with [OceanLight.jl](http://oceanlight.jl/). We’ll keep you posted on our progress.
