### 2025-09-20 haoboatlab/OceanLight.jl#11 @tayT0T
Thank you for the suggestion. We’ll be sure to use PRs starting with the next issue and from now on.
