### 2025-09-24 32d6d7a
Adding rough outline MPI tutorial page
Adding a section for MPI tutorial page. And visualize code.
