+++ docs/src/QuickStart/MPI.md
@@ -37,19 +37,160 @@ MPI.Init()
 ```
 
 ```@example MPI_tutorial 
+comm = MPI.COMM_WORLD
+# myid is rank of each cpu (tagname corresponding to each cpu being used)
+myid = MPI.Comm_rank(comm)
+# ncpu is number of cpu used to run the program
+ncpu = MPI.Comm_size(comm)
+```
+```@example MPI_tutorial 
+# the total number of photons in this case is the photons at each grid poin multiply by every single grid
+nind=p.nphoton
+# dind is the number photons that each cpu need to work on
 
+if mod(nind,ncpu) ==0
+    dind=div(nind,ncpu)
+else
+    dind=div(nind,ncpu) + 1
+end
+inds=myid*dind+1
+# inde is the total number of the photons that cpu would simulate
+inde=(myid+1)*dind
+# if total number is larger than the actual photons we want to calculate, set it equal
+
+if inde > nind
+    inde=nind
+end
+
+allind=1:p.nphoton
+"η is the height(z axis) corresponding to each grid point on the surface, 2d array of grid number in \
+x and y direction"
+η=zeros(p.nxs,p.nys)
+"ηx is the x coordination corresponding to each grid point on the surface, 2d array of grid number in\
+ x and y direction"
+ηx=zeros(p.nxs,p.nys)
+"ηy is the y coordination corresponding to each grid point on the surface, 2d array of grid number in\
+ x and y direction"
+ηy=zeros(p.nxs,p.nys)
+randrng = MersenneTwister(1234+myid)
+ϕps,θps=phasePetzold()
+
+ed=zeros(p.nx,p.ny,p.nz)
+esol=zeros(p.num,p.nz)
+area=zeros(4)
+interi=zeros(Int64,4)
+interj=zeros(Int64,4)
+ix=div(p.nxη,2)+1
+iy=div(p.nyη,2)+1
+
+
+η.=0
+ηx.=0
+ηy.=0
+ed.=0
+esol.=0
+MPI.Allreduce!(η,+,comm)
+MPI.Allreduce!(ηx,+,comm)
+MPI.Allreduce!(ηy,+,comm)
+if p.z[1] <= maximum(η)
+    error("ztop smaller than maximum η!")
+end
+println("myid $myid wave surface input complete!")
+xpb,ypb,zpb,θ,ϕ,fres=interface(η,ηx,ηy,p)
+println("myid $myid refraction at interface complete!")
+
+for ind=inds:inde
+    ip=allind[ind]
+    transfer!(ed,esol,θ[ix,iy],ϕ[ix,iy],fres[ix,iy],ip,
+                xpb[ix,iy],ypb[ix,iy],zpb[ix,iy],area,interi,interj,
+                                    randrng,η,ϕps,θps,p,1)
+end
+
+
+MPI.Allreduce!(ed,+,comm)
 ```
 
+```
+if myid ==0
+    applybc!(ed,p)
+    exported(ed,η,p,"rawdat/case$(icase)/ed","3D",176)
+end
+```
+```@example MPI_tutorial 
+MPI.Barrier(comm)
+MPI.Finalize()
+```
 
-
-
-
-
-
-
+# Run .jl file
 
 ```
 mpirun -np 4 julia FILE_NAME.jl
 ```
 
+# Visualization 
 
+
+
+
+
+
+
+```@example MPI_tutorial 
+using Plots
+using Plots.Measures
+
+# Choose slice indices
[…]
