+++ docs/src/QuickStart/MPI.md
@@ -1,5 +1,9 @@
 # Implementing MPI
 
+Since `OceanLight.jl` implements the Monte Carlo ray tracing algorithm, the accuracy of the solution depends on its convergence. However, this can be inefficient and time-consuming for specific problems that require a large number of photons `nphoton`. To address this issue, rather than running the entire simulation on a single CPU, we can allocate smaller chunks of photons across multiple CPUs, simulate them independently, and later combine the results into a single solution. That’s where MPI comes in.
+
+MPI, or Message Passing Interface, is a communication library designed for sending and receiving messages across multiple processors, making it suitable for parallel programming. To install MPI on your system, check out the [Installing Open MPI guide](https://docs.open-mpi.org/en/v5.0.x/installing-open-mpi/quickstart.html)
+
 ```@setup  MPI_tutorial
 cd(mktempdir()) 
 using Pkg 
@@ -9,6 +13,10 @@ Pkg.add("MPI")
 ```
 
 ## Problem
+
+In this example, the problem is to calculate the downwelling irradiance field when the surface is completely flat, and a total of 10,000,000 photons are focused at a single point in the center. Our domain of interest is defined as $x,y \in \left[\mathrm{-10m},\mathrm{10m}\right]$, and $z \in \left[\mathrm{-190m},\mathrm{10m}\right]$ in depth, corresponding to a grid resolution of $512 \times 512 \times 200$ points. Periodic boundary conditions are applied at the domain boundaries. The attenuation properties of water are characterized by an absorption coefficient of $a = 0.0196$ and a scattering coefficient of $b = 0.0031$, which correspond to the optical properties of seawater at a wavelength of $490 ,\mathrm{nm}$ [^1].
+
+To execute parallel programming using MPI, our Julia program must be run through a `.jl` script file. Inside the script file, we first create input data structures for the simulation. Here, `OceanLight.writeparams` will create a new input file in `.yml` format.
 
 ```@example MPI_tutorial 
 using OceanLight 
@@ -29,99 +37,106 @@ data=Dict("irradiance"=>Dict("nxe"=>nxe,"nye"=>nye,"nz"=>nz,"dz"=>dz,"ztop"=>zto
 OceanLight.writeparams(data)
 ```
 
-## Initialize MPI
+For a step-by-step description and the utility of each function, check out this [tutorial](https://haoboatlab.github.io/OceanLight.jl/dev/QuickStart/Center/). 
+
+## Initialize MPI and variables
+
+To enable parallel programming in Julia, we first import the `MPI` package and then initialize the environment with `MPI.Init()`.  
 
 ```@example MPI_tutorial 
 using MPI
 MPI.Init()
 ```
 
+First, `MPI.COMM_WORLD` is a predefined intracommunicator that represents the group of processes launched with MPI. We will declare it as the variable `comm`. Each processor inside the communicator has a unique rank, or ID number, which we can access through `MPI.Comm_rank(comm)`. Lastly, we can obtain the number of processors inside the communicator with `MPI.Comm_size(comm)`. 
+
 ```@example MPI_tutorial 
 comm = MPI.COMM_WORLD
-# myid is rank of each cpu (tagname corresponding to each cpu being used)
-myid = MPI.Comm_rank(comm)
-# ncpu is number of cpu used to run the program
-ncpu = MPI.Comm_size(comm)
[…]
