### 2025-09-29 0621fa2
Fix Typo in plotting
Fix typo in MPI.md plotting block section.
