### 2025-10-02 a367e75
Add similar packages in README and docs/src/index
Adding a paragraph about the similar packages, their utilities and differences, and also update reference in README.md file and docs/src/index.md
