+++ README.md
@@ -9,6 +9,8 @@
 
 Optical oceanography concerns all aspects of light and its interaction with seawater, which are crucial for addressing problems related to physical, biological, and chemical oceanographic processes, such as phytoplankton photosynthesis, biogeochemical cycles, and climate change [^3] [^4]. However, due to the complex interaction between light and free-surface wave geometry, the irradiance distribution can be highly variable [^5] [^6], making it difficult to obtain analytical solutions. **OceanLight.jl** enables physics-based, reproducible simulations of light distribution beneath the ocean surface.
 
+**OceanLight.jl** implements a Monte Carlo ray-tracing algorithm, whose accuracy depends on the number of simulated photon paths. This approach yields not only averaged quantities, such as mean intensity and flux, but also detailed three-dimensional distributions of irradiance intensity. Combined with its built-in light refraction calculator, OceanLight.jl is particularly well suited for modeling irradiance fields modulated by complex ocean surface geometries. An existing open-source package, OceanBioME.jl [^7], also includes a model of light attenuation caused by biomass, though it does not account for the modulation effect of fluctuating ocean surface waves. For more general problems that require only averaged radiative quantities, deterministic solvers of the radiative transfer equation (RTE) are available, such as PythonicDISORT [^8] and DISORT [^9].
+
 ## Installation 
 
 OceanLight requires Julia software. To do so, 
@@ -191,17 +193,22 @@ For more information, check out our [contributor guide](https://haoboatlab.githu
 <img  src="https://raw.githubusercontent.com/haoboatlab/OceanLight.jl/main/docs/src/assets/Wholegrid1000.png" width="600" align="center">
 *Simulation of 1000 Photons at the every grid point with observed surface elevation. (a) irradiance field on the horizontal plane at 30 m depth. (b) irradiance field on the horizontal plane at 150 m depth. (c) irradiance field on the vertical plane at the center.*
 
-
 ## Reference 
 
 [^1]: Hao, X., & Shen, L. (2022). A novel machine learning method for accelerated modeling of the downwelling irradiance field in the upper ocean. Geophysical Research Letters, 49, e2022GL097769. https://doi.org/10.1029/2022GL097769
 
 [^2]: Kirk, J. T. O. (1981). Monte Carlo procedure for simulating the penetration of light into natural waters. In Technical paper - Commonwealth Scientific and Industrial Research Organization (Vol. 36).
 
-[^3]: Dickey, T. D., Kattawar, G. W., & Voss, K. J. (2011). Shedding new light on light in the ocean. Physics Today, 64(4), 44-49.
+[^3]: Dickey, T. D., Kattawar, G. W., & Voss, K. J. (2011). Shedding new light on light in the ocean. Physics Today, 64(4), 44-49. https://doi.org/10.1063/1.3580492
 
-[^4]: Dickey, T., Lewis, M., & Chang, G. (2006). Optical oceanography: recent advances and future directions using global remote sensing and in situ observations. Reviews of geophysics, 44(1).
+[^4]: Dickey, T., Lewis, M., & Chang, G. (2006). Optical oceanography: recent advances and future directions using global remote sensing and in situ observations. Reviews of geophysics, 44(1). https://doi.org/10.1029/2003RG000148
 
-[^5]: Darecki, M., Stramski, D., & Sokólski, M. (2011). Measurements of high‐frequency light fluctuations induced by sea surface waves with an Underwater Porcupine Radiometer System. Journal of Geophysical Research: Oceans, 116(C7).
[…]
