### 2025-10-09 a83864d
Fix grammar/ Make MPI line less ambigous in JOSS.md
Fixing the grammar, and rewrite the sentence related the use of MPI with OceanLight.jl. Restate that OceanLight can be used along with MPI to accomodate parallel programming.
