### 2025-10-14 d6a4baa
Revise similar package section in README and index.md
Revise similar package section under the Overview subtopic in README.md and docs/src/index.md. Adding the standalone package of OceanLight.jl and the plane-parallel approximation for OceanBioME.jl
