### 2025-10-14 ef749b4
Revising and fixing grammar in MPI.md
