### 2025-10-14 45394b8
Fixing variable error in plotting section MPI.md
Fixing Variable error that require the struct param p in the front, in
the plotting section in docs/src/QuickStart/MPI.md
