### 2025-10-23 49c0046
Add flowchart images in MPI.md
Adding a flowchart images in docs/src/Quickstart/MPI.md, to help illustrating the use of MPI.
