### 2025-10-24 #8537 @glwagner
There is a still a small issue in that the package authors include what I regard as misleading documentation about MPI. In fact, the package does not use MPI. While it is true that MPI can still be used to launch multiple processes (as it can for all code), there is no specific need to do that with the package (there are better alternatives for parallelization, as far as I can tell).
