### 2025-10-27 d993bd6
Remove MPI from tutorial page, JOSS paper
Remove MPI tutorial page from QuickStart. Also not mentioning the MPI in the JOSS paper.
