+++ joss/paper.md
@@ -33,7 +33,7 @@ Optical oceanography concerns all aspects of light and its interaction with seaw
 
 One prominent numerical method is the implementation of the MC approach to approximate the light field. The general idea is to construct probability distributions of photon trajectories by repeatedly sampling random paths, including scattering, absorption, and path lengths. Despite its generality, this method has proven to be powerful [@Mobley:1994]. The MC method has been widely applied to various specific problems related to the calculation of downward irradiance field. For instance, an MC radiative transfer solver has been developed that accounts for the shadowing effect and detector size [@Xu:2014], and direct MC simulations have been performed for both polarized and unpolarized light [@Xu:2011].   
 
-Here, we introduce `OceanLight`, a Julia-based software package designed to facilitate three-dimensional direct MC simulation of light fields in seawater. `OceanLight` is based on the forward MC algorithm as described by @Mobley:1994. It includes light refraction at the air-water interface, where the refraction angle in the water phase is computed using Snell's Law, and the proportion of transmitted light is determined by the Fresnel equations. The probability distribution of the polar angles in scattering events is based on the measurements reported by @Petzold:1972 [@Kirk:1981]. Additionally, `OceanLight` supports parallel computing to enhance performance and reduce computational time. The downwelling irradiance fields generated by `OceanLight` also serves as a valueable resource for machine learning applications, enabling the development of accelerated models that require lower computational costs [@Hao:2022]. 
+Here, we introduce `OceanLight`, a Julia-based software package designed to facilitate three-dimensional direct MC simulation of light fields in seawater. `OceanLight` is based on the forward MC algorithm as described by @Mobley:1994. It includes light refraction at the air-water interface, where the refraction angle in the water phase is computed using Snell's Law, and the proportion of transmitted light is determined by the Fresnel equations. The probability distribution of the polar angles in scattering events is based on the measurements reported by @Petzold:1972 [@Kirk:1981]. The downwelling irradiance fields generated by `OceanLight` also serves as a valueable resource for machine learning applications, enabling the development of accelerated models that require lower computational costs [@Hao:2022]. 
 
 # Main Features
 
@@ -51,7 +51,7 @@ The main features of `OceanLight` are illustrated in figure 2.
 
 At the air–water interface, `OceanLight.interface` calculates the polar angle $\theta$ and azimuthal angle $\phi$ at which photons are transmitted into the water, using Snell’s Law. The portions of transmitted and reflected light energy are determined using the Fresnel equations.
 
-Once photons are transmitted into the water, `OceanLight.transfer` simulates their paths using the direct MC method. Each photon is tracked as it travels through the water column, with its position updated after every scattering or absorption event. The process continues until the photon either vanishes in absorption or reaches the bottom boundary. To accelerate computation, `OceanLight` can be integrated with MPI to enable parallel processing.
+Once photons are transmitted into the water, `OceanLight.transfer` simulates their paths using the direct MC method. Each photon is tracked as it travels through the water column, with its position updated after every scattering or absorption event. The process continues until the photon either vanishes in absorption or reaches the bottom boundary.
 
 Simulation output can be exported in HDF5 (`.h5`) format using `OceanLight.exported`, allowing users access to the full three-dimensional irradiance field and its statistical properties.
 
