+++ docs/make.jl
@@ -5,8 +5,7 @@ makedocs(
     pages = [
         "Home" => "index.md",
         "Quick Start" => [
-        "Photons at the center" => "QuickStart/Center.md",
-        "Implement MPI" => "QuickStart/MPI.md"
+        "Photons at the center" => "QuickStart/Center.md"
         ],
         "Module" => [
         "Parameters" => "module/Parameters.md",
