+++ docs/src/QuickStart/MPI.md
@@ -1,240 +1,0 @@
-# Implementing MPI
-
-Since `OceanLight.jl` implements the Monte Carlo ray tracing algorithm, the accuracy of its solution depends on convergence. However, this process can be inefficient and time-consuming for problems that require a large number of photons `N`. To address this issue, instead of running the entire simulation on a single CPU, we can divide the workload by allocating smaller chunks of photons across multiple CPUs, simulate them independently, and then combine the results into a single solution. This parallelization reduces the computational cost from $\mathcal{O}(N)$ for a single CPU to $\mathcal{O}(N/n)$ when using $n$ CPUs.
-
-![MPI_flowchart](https://raw.githubusercontent.com/haoboatlab/OceanLight.jl/main/docs/src/assets/mpi_flowchart.png)
-
-*Flowchart and computational cost comparison between single CPU and multiple CPUs. Here, $N$ denotes the total number of photons and $I_0$ represents the resulting irradiance field.*
-
-MPI, or Message Passing Interface, is a communication library designed for sending and receiving messages across multiple processors, making it suitable for parallel programming. To install MPI on your system, check out the [Installing Open MPI guide](https://docs.open-mpi.org/en/v5.0.x/installing-open-mpi/quickstart.html)
-
-```@setup  MPI_tutorial
-cd(mktempdir()) 
-using Pkg 
-Pkg.add("OceanLight") 
-Pkg.add("Plots") 
-Pkg.add("MPI")
-```
-
-## Problem
-
-In this example, the goal is to calculate the downwelling irradiance field when the surface is completely flat, and a total of 10,000,000 photons are focused at a single point at the center. The domain of interest is defined as $x,y \in \left[\mathrm{-10m},\mathrm{10m}\right]$, and $z \in \left[\mathrm{-190m},\mathrm{10m}\right]$ in depth, corresponding to a grid resolution of $512 \times 512 \times 200$ points. Periodic boundary conditions are applied at the domain boundaries. The attenuation properties of water are characterized by an absorption coefficient of $a = 0.0196$ and a scattering coefficient of $b = 0.0031$, which correspond to the optical properties of seawater at a wavelength of $490 ,\mathrm{nm}$ [^1].
-
-To execute parallel programming using MPI, the Julia program must be run from a `.jl` script file. Inside the script, we first create the input data structures for the simulation. Here, `OceanLight.writeparams` generates a new input file in `.yml` format.
-
-```@example MPI_tutorial 
-using OceanLight 
-using Random
-```
-```@example MPI_tutorial 
-# irradiance
-nz = 200; dz = 1; nxe = 512; nye = 512; num = 31; ztop = 10                 
-# photon
-nphoton = 10000000; kr = 10; nxp = 512; kbc = 0; b = 0.0031; nyp = 512; a = 0.0196;                 
-# wave
-pey = 2*pi/20.0; nxeta = 512; nyeta = 512; pex = 2*pi/20.0;
-
-data=Dict("irradiance"=>Dict("nxe"=>nxe,"nye"=>nye,"nz"=>nz,"dz"=>dz,"ztop"=>ztop,"num"=>num),
-            "wave"=>Dict("pex"=>pex,"pey"=>pey,"nxeta"=>nxeta,"nyeta"=>nyeta),
[…]
