### 2025-11-01 e944386
Expand abbreviation for Monte Carlo method
I think that we should assume the abstract/summary as a separate document and redefine abbreviations and acronyms again in the main text.
