+++ joss/paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 
 Optical oceanography concerns all aspects of light and its interaction with seawater, which are crucial for addressing problems related to physical, biological, and chemical oceanographic processes, such as phytoplankton photosynthesis, biogeochemical cycles, and climate change [@Dickey:2011;@Dickey:2006]. However, due to the complex interaction between light and free-surface wave geometry, the irradiance distribution can be highly variable [@Darecki:2011;@Gernez:2011], making it difficult to obtain analytical solutions.
 
-One prominent numerical method is the implementation of the MC approach to approximate the light field. The general idea is to construct probability distributions of photon trajectories by repeatedly sampling random paths, including scattering, absorption, and path lengths. Despite its generality, this method has proven to be powerful [@Mobley:1994]. The MC method has been widely applied to various specific problems related to the calculation of downward irradiance field. For instance, an MC radiative transfer solver has been developed that accounts for the shadowing effect and detector size [@Xu:2014], and direct MC simulations have been performed for both polarized and unpolarized light [@Xu:2011].   
+One prominent numerical method is the implementation of the Monte Carlo (MC) approach to approximate the light field. The general idea is to construct probability distributions of photon trajectories by repeatedly sampling random paths, including scattering, absorption, and path lengths. Despite its generality, this method has proven to be powerful [@Mobley:1994]. The MC method has been widely applied to various specific problems related to the calculation of downward irradiance field. For instance, an MC radiative transfer solver has been developed that accounts for the shadowing effect and detector size [@Xu:2014], and direct MC simulations have been performed for both polarized and unpolarized light [@Xu:2011].   
 
 Here, we introduce `OceanLight`, a Julia-based software package designed to facilitate three-dimensional direct MC simulation of light fields in seawater. `OceanLight` is based on the forward MC algorithm as described by @Mobley:1994. It includes light refraction at the air-water interface, where the refraction angle in the water phase is computed using Snell's Law, and the proportion of transmitted light is determined by the Fresnel equations. The probability distribution of the polar angles in scattering events is based on the measurements reported by @Petzold:1972 [@Kirk:1981]. Additionally, `OceanLight` supports parallel computing to enhance performance and reduce computational time. The downwelling irradiance fields generated by `OceanLight` also serves as a valueable resource for machine learning applications, enabling the development of accelerated models that require lower computational costs [@Hao:2022]. 
 
