### 2025-11-01 019d6b4
Revise citation format
There were a missing @
