### 2025-11-01 haoboatlab/HydrOptics.jl#19 @iuryt
OceanLight.jl or WaterLight.jl?
Perhaps I am wrong, but so far, I have not yet understood why this package is only applied to seawater and not to water bodies in general. I think that would significantly expand the community of potential users, since this is also important to limnology. Perhaps consider calling it WaterLight.jl instead?
If I understand correctly, that would only change the parameters?

xref: https://github.com/openjournals/joss-reviews/issues/8537
