### 2025-11-04 haoboatlab/OceanLight.jl#19 @iuryt
I like `HydroOptics.jl`. Any opinion about this @glwagner ?
