### 2025-11-04 haoboatlab/OceanLight.jl#19 @glwagner
Both are great --- very similar! You could try HydrOptics.jl :-D
