### 2025-11-05 12263f5
Renaming OceanLight.jl to HydrOptics.jl
In src/ folder, rename a file to HydrOptics.jl, change the module name to HydrOptics
In Project.toml, change name of package, uuid, and version
In test/runtests.jl, change everything to HydrOptics
In joss/paper.md, change everything to HydrOptics
In README.md, change the text in paragraph to HydrOptics,not the code block and link URL
In docs/make.jl, change everything and repo URL to HydrOptics
In docs/src, change the text in paragrapth to HydrOptics, not the code block and link URL
