+++ docs/src/QuickStart/Center.md
@@ -1,8 +1,8 @@
 # Quick Start 
 
-In this example, we introduce the `OceanLight` calculation of downwelling irradiance field underneath the flat surface. The code example here can be directly pasted onto Julia terminal, run through `.jl` script file, or IJulia notebook. 
+In this example, we introduce the `HydrOptics` calculation of downwelling irradiance field underneath the flat surface. The code example here can be directly pasted onto Julia terminal, run through `.jl` script file, or IJulia notebook. 
 
-First, we import `OceanLight` packages and another `random` dependent package, that will be used as a seed to generate random number in the Monte Carlo simulation.
+First, we import `HydrOptics` packages and another `random` dependent package, that will be used as a seed to generate random number in the Monte Carlo simulation.
 
 ```@setup  Center
 cd(mktempdir()) 
@@ -21,9 +21,9 @@ In this example, the problem is to calculate the downwelling irradiance field, w
 
 ##  Initial Condition 
 
-OceanLight accesses all input variables in `.yml` format and stores their values in `Param` structure.
+HydrOptics accesses all input variables in `.yml` format and stores their values in `Param` structure.
 
-All input variables required by OceanLight can be separated into 3 categories:
+All input variables required by HydrOptics can be separated into 3 categories:
 1. **Irradiance:** Setting up the dimension of downwelling irradiance field calculation grid. 
 2. **Photon:** Number of Photons and optical properties of the water.
 3. **Wave:** Structure of surface wave field. 
@@ -74,14 +74,14 @@ p = OceanLight.readparams()
 
 Before the simulation, user needs to declare and initialize all parameters and their dimensions. 
 
-During the light refraction between two mediums (air and water) calculation, `OceanLight` requires the surface elevation attribution: $\eta$; surface elevation, $\eta_{x}$; partial derivative of $\eta$ in x direction, and $\eta_{y}$; partial derivative of $\eta$ in y direction. All surface elevation distribution, as specified above, need to have the same dimension with the incoming photons' grid. Hence, 
+During the light refraction between two mediums (air and water) calculation, `HydrOptics` requires the surface elevation attribution: $\eta$; surface elevation, $\eta_{x}$; partial derivative of $\eta$ in x direction, and $\eta_{y}$; partial derivative of $\eta$ in y direction. All surface elevation distribution, as specified above, need to have the same dimension with the incoming photons' grid. Hence, 
 
 ```@example Center
 η = zeros(p.nxs,p.nys)
 ηx = zeros(p.nxs,p.nys)
 ηy = zeros(p.nxs,p.nys)
 ```
-After photons' interaction with the surface, OceanLight requires the information of specific coordinate of photon in cartesian grid ${xpb,ypb,zpb}$, the direction in which photon will travel in polar coordinate ${θ,ϕ}$, and the fraction of light that transmit through the water ${fres}$: all in the dimension of incoming photon grid size. 
+After photons' interaction with the surface, HydrOptics requires the information of specific coordinate of photon in cartesian grid ${xpb,ypb,zpb}$, the direction in which photon will travel in polar coordinate ${θ,ϕ}$, and the fraction of light that transmit through the water ${fres}$: all in the dimension of incoming photon grid size. 
 
 ```@example Center
 xpb = zeros(p.nxp,p.nyp);
@@ -92,9 +92,9 @@ zpb = zeros(p.nxp,p.nyp);
 fres = zeros(p.nxp,p.nyp);
 ```
 
-During the air-water interaction process, OceanLight simulates the photons transfer directly downward from the air side, interacts with the water surface, and transfer down into water medium. 
[…]
