+++ docs/src/index.md
@@ -1,10 +1,10 @@
-# OceanLight.jl Documentation
+# HydrOptics.jl Documentation
 
 ## Overview 
 
-**OceanLight.jl** calculates the downwelling irradiance field in the upper ocean. It uses the Monte Carlo method to simulate the trajectories of photons, from their refraction at the air–water interface to specific depths beneath the surface [^1] [^2]. 
+**HydrOptics.jl** calculates the downwelling irradiance field in the upper ocean. It uses the Monte Carlo method to simulate the trajectories of photons, from their refraction at the air–water interface to specific depths beneath the surface [^1] [^2]. 
 
-Optical oceanography concerns all aspects of light and its interaction with seawater, which are crucial for addressing problems related to physical, biological, and chemical oceanographic processes, such as phytoplankton photosynthesis, biogeochemical cycles, and climate change [^3] [^4]. However, due to the complex interaction between light and free-surface wave geometry, the irradiance distribution can be highly variable [^5] [^6], making it difficult to obtain analytical solutions. **OceanLight.jl** enables physics-based, reproducible simulations of light distribution beneath the ocean surface.
+Optical oceanography concerns all aspects of light and its interaction with seawater, which are crucial for addressing problems related to physical, biological, and chemical oceanographic processes, such as phytoplankton photosynthesis, biogeochemical cycles, and climate change [^3] [^4]. However, due to the complex interaction between light and free-surface wave geometry, the irradiance distribution can be highly variable [^5] [^6], making it difficult to obtain analytical solutions. **HydrOptics.jl** enables physics-based, reproducible simulations of light distribution beneath the ocean surface.
 
 ## Quick Install 
 
@@ -17,7 +17,7 @@ using Pkg
 Pkg.add("OceanLight")
 ```
 
-After installing, verify that OceanLight works as intended by:
+After installing, verify that HydrOptics works as intended by:
 
 ```Julia
 Pkg.test("OceanLight")
@@ -166,7 +166,7 @@ plot(p1, p2, p3, layout=l,
 
 ![Center1e7](https://raw.githubusercontent.com/haoboatlab/OceanLight.jl/main/docs/src/assets/center1e7.png)
 
-For a complete guide with details on each function and step, see the [OceanLight's Documentation](https://haoboatlab.github.io/OceanLight.jl/dev/QuickStart/Center/). 
+For a complete guide with details on each function and step, see the [HydrOptics's Documentation](https://haoboatlab.github.io/OceanLight.jl/dev/QuickStart/Center/). 
 
 ## Gallery
 

[…]
