### 2025-11-11 342576e
Change every "OceanLight" to HydrOptics
Change everything (URL link and function)
README.md, index.md, contribute.md, documentation, Project.toml for documentation.
