+++ docs/src/QuickStart/Center.md
@@ -7,11 +7,11 @@ First, we import `HydrOptics` packages and another `random` dependent package, t
 ```@setup  Center
 cd(mktempdir()) 
 using Pkg 
-Pkg.add("OceanLight") 
+Pkg.add("HydrOptics") 
 Pkg.add("Plots") 
 ```
 ```@example Center 
-using OceanLight 
+using HydrOptics 
 using Random
 ```
 
@@ -50,24 +50,24 @@ nxeta = 512                 # Number of surface elevation grid point in x direct
 nyeta = 512                 # Number of surface elevation grid point in y direction
 pex = 2*pi/20.0             # Lowest wavenumber that can be captured during the derivative of surface elevation in y direction
 ```
-**NOTE:** `num` must be set to a constant value of 31 (number of angle measurement in Kirk, 1981 [^2]). `kbc` accepts only binary values (0 or 1), representing periodic boundary conditions. In contrast to grid spacing `dz` in z-direction, the grid spacings in the x- and y-directions (`dx` and `dy`) are calculated automatically by `OceanLight.readparams`,  using the formulas $dx \times nxe  = \frac{2\pi}{pex}$, and $dy \times nye = \frac{2\pi}{pey}$. For more detail on all parameters  used, see [`Simulation parameters`](@ref simulation_parameters). 
+**NOTE:** `num` must be set to a constant value of 31 (number of angle measurement in Kirk, 1981 [^2]). `kbc` accepts only binary values (0 or 1), representing periodic boundary conditions. In contrast to grid spacing `dz` in z-direction, the grid spacings in the x- and y-directions (`dx` and `dy`) are calculated automatically by `HydrOptics.readparams`,  using the formulas $dx \times nxe  = \frac{2\pi}{pex}$, and $dy \times nye = \frac{2\pi}{pey}$. For more detail on all parameters  used, see [`Simulation parameters`](@ref simulation_parameters). 
 
-To create a input variable file suitable for this package, user can either create a new file in `.yml` format, copy, and paste the code block above, or using a build-in function `OceanLight.writeparams` to automate the process. 
+To create a input variable file suitable for this package, user can either create a new file in `.yml` format, copy, and paste the code block above, or using a build-in function `HydrOptics.writeparams` to automate the process. 
 
-The function `OceanLight.writeparams` converts the dictionary of input variables into the `light.yml` file. 
+The function `HydrOptics.writeparams` converts the dictionary of input variables into the `light.yml` file. 
 
 ```@example Center
 data=Dict("irradiance"=>Dict("nxe"=>nxe,"nye"=>nye,"nz"=>nz,"dz"=>dz,"ztop"=>ztop,"num"=>num),
             "wave"=>Dict("pex"=>pex,"pey"=>pey,"nxeta"=>nxeta,"nyeta"=>nyeta),
             "photon"=>Dict("nxp"=>nxp,"nyp"=>nyp,"nphoton"=>nphoton,"a"=>a,"b"=>b,"kr"=>kr,"kbc"=>kbc))
 
-OceanLight.writeparams(data)
+HydrOptics.writeparams(data)
 ```
 
-Once the input variable file is generated, this package calculates other related variables and stores all values in the `Param` structure, through `OceanLight.readparams`. The default input variable file name that will be read is `light.yml`, but can be specified by users. In this case, we store `Param` structure in variable `p`. 
[…]
