+++ README.md
@@ -1,6 +1,6 @@
 # HydrOptics.jl
 
-[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://haoboatlab.github.io/OceanLight.jl/dev/)
+[![Dev](https://img.shields.io/badge/docs-dev-blue.svg)](https://haoboatlab.github.io/HydrOptics.jl/dev/)
 [![License](https://img.shields.io/badge/License-MIT-blue.svg?style=flat-square)](https://mit-license.org/)
 
 ## Overview
@@ -19,13 +19,13 @@ HydrOptics requires Julia software. To do so,
 
 ```Julia
 using Pkg
-Pkg.add("OceanLight")
+Pkg.add("HydrOptics")
 ```
 
 After installing, verify that HydrOptics works as intended by:
 
 ```Julia
-Pkg.test("OceanLight")
+Pkg.test("HydrOptics")
 ```
 
 ## Running your first model 
@@ -33,7 +33,7 @@ Pkg.test("OceanLight")
 As a simple example, let’s calculate the downwelling irradiance field when the surface is completely flat and a total of 10,000,000 photons are focused at a single point in the center.
 
 ```Julia
-using OceanLight 
+using HydrOptics 
 using Random
 
 # irradiance
@@ -61,8 +61,8 @@ data=Dict("irradiance"=>Dict("nxe"=>nxe,"nye"=>nye,"nz"=>nz,"dz"=>dz,"ztop"=>zto
             "wave"=>Dict("pex"=>pex,"pey"=>pey,"nxeta"=>nxeta,"nyeta"=>nyeta),
             "photon"=>Dict("nxp"=>nxp,"nyp"=>nyp,"nphoton"=>nphoton,"a"=>a,"b"=>b,"kr"=>kr,"kbc"=>kbc))
 
-OceanLight.writeparams(data)
-p = OceanLight.readparams()
+HydrOptics.writeparams(data)
+p = HydrOptics.readparams()
 
 η = zeros(p.nxs,p.nys)
 ηx = zeros(p.nxs,p.nys)
@@ -81,14 +81,14 @@ interi = zeros(Int64,4)
 interj = zeros(Int64,4)
 ix = div(p.nxη,2)+1
 iy = div(p.nyη,2)+1
-ϕps,θps = OceanLight.phasePetzold()
+ϕps,θps = HydrOptics.phasePetzold()
 
-OceanLight.interface!(xpb,ypb,zpb,θ,ϕ,fres,η,ηx,ηy,p)
+HydrOptics.interface!(xpb,ypb,zpb,θ,ϕ,fres,η,ηx,ηy,p)
 for ip = 1:p.nphoton
-    OceanLight.transfer!(ed,esol,θ[ix,iy],ϕ[ix,iy],fres[ix,iy],ip,xpb[ix,iy],
+    HydrOptics.transfer!(ed,esol,θ[ix,iy],ϕ[ix,iy],fres[ix,iy],ip,xpb[ix,iy],
         ypb[ix,iy],zpb[ix,iy],area,interi,interj,randrng,η,ϕps,θps,p,1)
 end
-OceanLight.applybc!(ed,p)
+HydrOptics.applybc!(ed,p)
 
 max_val, max_loc = findmax(ed)
 ed = ed./max_val
@@ -170,25 +170,25 @@ plot(p1, p2, p3, layout=l,
      left_margin=10mm, right_margin=10mm)
 ```
 </details>
-<img  src="https://raw.githubusercontent.com/haoboatlab/OceanLight.jl/main/docs/src/assets/center1e7.png" width="600" align="center">
+<img  src="https://raw.githubusercontent.com/haoboatlab/HydrOptics.jl/main/docs/src/assets/center1e7.png" width="600" align="center">
 
[…]

[…]
