+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"QuickStart/Center/#Quick-Start","page":"Photons at the center","title":"Quick Start","text":"","category":"section"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"In this example, we introduce the HydrOptics calculation of downwelling irradiance field underneath the flat surface. The code example here can be directly pasted onto Julia terminal, run through .jl script file, or IJulia notebook. ","category":"page"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"First, we import HydrOptics packages and another random dependent package, that will be used as a seed to generate random number in the Monte Carlo simulation.","category":"page"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"using OceanLight \nusing Random","category":"page"},{"location":"QuickStart/Center/#Problem","page":"Photons at the center","title":"Problem","text":"","category":"section"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"In this example, the problem is to calculate the downwelling irradiance field, when the surface is completely flat, and a total of 10,000,000 photons is focused at a single point in the center. Our domain of interest is defined as xy in leftmathrm-10mmathrm10mright, and z in leftmathrm-190mmathrm10mright in depth, corresponding to a grid resolution of 512 times 512 times 200  points. Periodic boundary conditions are applied at the domain boundaries. The attenuation properties of water are characterized by an absorption coefficient of a = 00196 and scattering coefficient b = 00031, which corresponding to the optical properties of sea water at wavelength 490 mathrmnm [1]. ","category":"page"},{"location":"QuickStart/Center/#Initial-Condition","page":"Photons at the center","title":"Initial Condition","text":"","category":"section"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"HydrOptics accesses all input variables in .yml format and stores their values in Param structure.","category":"page"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"All input variables required by HydrOptics can be separated into 3 categories:","category":"page"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"Irradiance: Setting up the dimension of downwelling irradiance field calculation grid. \nPhoton: Number of Photons and optical properties of the water.\nWave: Structure of surface wave field. ","category":"page"},{"location":"QuickStart/Center/","page":"Photons at the center","title":"Photons at the center","text":"# irradiance\nnz = 200                    # Number of total grid point in z direction\ndz = 1                      # Physical length of grid spacing in z direction\nnxe = 512                   # Number of grid point of the calculation grid in x direction  \nnye = 512                   # Number of grid point of the calculation grid in y direction  \nnum = 31                    # Constant value (number of angle measurement in Kirk,1981)\nztop = 10                   # Number of grid point in air phase in z direction\n# photon\nnphoton = 10000000          # Number of photons being generated at each grid point \nkr = 10                     # Dummy variable (in developement, not bein
[…]
