### 2025-11-11 02d6651
fixing typo Readme, index.md
