### 2025-11-12 0d23759
Fix flowchart image to HydrOptics
In joss paper, fixing the flowchart image from previous name "OceanLight" to "HydrOptics"
