### 2025-11-12 #8537 @haoboatlab
Yes. All package registration, documentation, tutorials, and the JOSS paper have been updated accordingly.
