### 2025-11-14 4ed8c5f
joss/docs Adjust title and statement of need
Make the title and statment of need more generic (Change Ocean to water body)
