+++ joss/paper.md
@@ -1,5 +1,5 @@
 ---
-title: "HydrOptics.jl: A Julia package for calculating irradiance in the upper ocean using Monte Carlo simulations"
+title: "HydrOptics.jl: A Julia package for calculating irradiance in the water body using Monte Carlo simulations"
 tags:
   - Julia
   - Ocean
@@ -33,7 +33,7 @@ Optical oceanography concerns all aspects of light and its interaction with seaw
 
 One prominent numerical method is the implementation of the Monte Carlo (MC) approach to approximate the light field. The general idea is to construct probability distributions of photon trajectories by repeatedly sampling random paths, including scattering, absorption, and path lengths. Despite its generality, this method has proven to be powerful [@Mobley:1994]. The MC method has been widely applied to various specific problems related to the calculation of downward irradiance field. For instance, an MC radiative transfer solver has been developed that accounts for the shadowing effect and detector size [@Xu:2014], and direct MC simulations have been performed for both polarized and unpolarized light [@Xu:2011].   
 
-Here, we introduce `HydrOptics`, a Julia-based software package designed to facilitate three-dimensional direct MC simulation of light fields in seawater. `HydrOptics` is based on the forward MC algorithm as described by @Mobley:1994. It includes light refraction at the air-water interface, where the refraction angle in the water phase is computed using Snell's Law, and the proportion of transmitted light is determined by the Fresnel equations. The probability distribution of the polar angles in scattering events is based on the measurements reported by @Petzold:1972 [@Kirk:1981]. The downwelling irradiance fields generated by `HydrOptics` also serves as a valueable resource for machine learning applications, enabling the development of accelerated models that require lower computational costs [@Hao:2022]. 
+Here, we introduce `HydrOptics`, a Julia-based software package designed to facilitate three-dimensional direct MC simulation of light fields in the water body. `HydrOptics` is based on the forward MC algorithm as described by @Mobley:1994. It includes light refraction at the air-water interface, where the refraction angle in the water phase is computed using Snell's Law, and the proportion of transmitted light is determined by the Fresnel equations. The probability distribution of the polar angles in scattering events is based on the measurements reported by @Petzold:1972 [@Kirk:1981]. The downwelling irradiance fields generated by `HydrOptics` also serves as a valueable resource for machine learning applications, enabling the development of accelerated models that require lower computational costs [@Hao:2022]. 
 
 # Main Features
 
@@ -45,7 +45,7 @@ The main features of `HydrOptics` are illustrated in figure 2.
 
 1. **Irradiance:** resolution of the solution grid, attenuation coefficients, and boundary condition specification 
 2. **Photon:** number of photons and grid spacing
-3. **Wave:** ocean surface elevation and related attributes 
+3. **Wave:** surface elevation and related attributes 
 
 `HydrOptics` reads input parameters from `.yml` files and stores them in Julia's `struct` format. In addition, surface wave attributes, comprising surface elevation and its partial derivatives in the x and y directions, are required.  Users can either generate randomized surface elevations using `HydrOptics.setwave!` or import their own data and convert it to a format that aligns with the photon entry grid using `HydrOptics.convertwave!`.
 
