### 2025-11-15 071ebfb
Update README.md with Zenodo badge
