+++ v2.0.1/QuickStart/light.yml
@@ -1,0 +1,20 @@
+irradiance:
+  nz: 200
+  dz: 1
+  nxe: 512
+  nye: 512
+  num: 31
+  ztop: 10
+photon:
+  nphoton: 10000000
+  kr: 10
+  nxp: 512
+  kbc: 0
+  b: 0.0031
+  nyp: 512
+  a: 0.0196
+wave:
+  pey: 0.3141592653589793
+  nxeta: 512
+  nyeta: 512
+  pex: 0.3141592653589793
