### 2025-11-18 9f1a1c4
Fix image label in joss/paper.md
