### 2026-01-30 #8537 @haoboatlab
Hi @AnjaliSandip I wanted to follow up on the status of our manuscript. I believe we have fixed all the issues upon your recommendation of acceptance. Could you provide a guidance on what is needed, if any, from our side? If not, I'm hoping that it can get published ASAP. Thank you.
