### 2026-02-01 #8537 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
